+++ src/js/components/IssuanceResultSection.js
@@ -1,4 +1,4 @@
-import { $, $$ } from "../dom.js";
+import { $, $$ } from "../utils/dom.js";
 
 export default class IssuanceResultSection {
   constructor() {
