+++ src/js/components/Modal.js
@@ -1,4 +1,4 @@
-import { $ } from "../dom.js";
+import { $ } from "../utils/dom.js";
 
 export default class Modal {
   constructor() {
